refactor(vscode): move commands and findBinary to own files
